move classes from generator.cs to its own file
